+++ src/changelog.h
@@ -1279,22 +1279,33 @@
 // 03.01.05   JDM - Aug 30, 2024    - Defect repair, minor cleanup:
 //                                      - Related to issue #502: added HG to allowed RSG stellar type check, preventing GB winds from being applied during HG.
 //                                      - Changed all "FLEXIBLE2023" naming to "MERRITT2024"
-// 03.01.06   JR - Aug 28, 2024     - Enhancement:
+// 03.01.06   JR - Aug 30, 2024     - Enhancement:
 //                                      - Added functionality to allow users to specify if WD-binaries should be included in the BSE DCO file
 //                                        New option: --include-WD-binaries-as-DCO
 //                                        See "What's New" and option documentation for details.
 //                                        (Issue #1170)
 //                                      - added deprecation notice for '--mass-loss-prescription NONE' (should use ZERO) - missed in v03.00.00
-// 03.02.00   SS - Sep 4, 2024      - Enhancement:
+// 03.01.07   JDM - Sep 05, 2024    - Defect repair:
+//                                      - Set wind mass loss for remnants to zero. 
+// 03.01.08   JR - Sep 06, 2024     - Defect repair, typo fixes:
+//                                      - Fix for issue #1219: Option --black-hole-kicks-mode (aka --black-hole-kicks) ignored
+//                                      - Fixed some stray typos
+// 03.01.09   IM - Sep 07, 2024     - Defect repair:
+//                                      - Fix for issue #1218: increased default MULLERMANDEL_REMNANT_MASS_MAX_ITERATIONS, but on failure to find a solution,
+//                                          indicating a narrow range, just pick a midpoint; remove associated error
+//                                      - Corrected --mass-loss-prescription description in documentation
+// 03.02.00   SS - Sep 11, 2024      - Enhancement:
 //                                      - Improvements to modelling of chemically homogeneous evolution
 //                                      - New options: --enable-rotationally-enhanced-mass-loss,
 //                                        --enhance-CHE-lifetimes-luminosities, --scale-CHE-mass-loss-with-surface-helium-abundance,
 //                                        --scale-terminal-wind-velocity-with-metallicity-power
 //                                      - To facilitate --scale-CHE-mass-loss-with-surface-helium-abundance, added basic tracking of 
-//                                        surface and core hydrogen and helium abundances for CH stars.
+//                                        surface and core hydrogen and helium abundances.
 //                                        See "What's New" and option documentation for details
 //
 
+
 const std::string VERSION_STRING = "03.02.00";
+
 
 # endif // __changelog_h__