+++ src/constants.h
@@ -269,7 +269,7 @@ constexpr int    ADAPTIVE_MASS0_MAX_TRIES               = 30;                   
 constexpr int    ADAPTIVE_MASS0_MAX_ITERATIONS          = 50;                                                       // Maximum number of iterations in HG::Mass0ToMatchDesiredCoreMass()
 constexpr double ADAPTIVE_MASS0_SEARCH_FACTOR_FRAC      = 1.0;                                                      // Search size factor (fractional part) in HG::Mass0ToMatchDesiredCoreMass() (added to 1.0)
 
-constexpr int    MULLERMANDEL_REMNANT_MASS_MAX_ITERATIONS = 100;                                                    // Maximum number of iterations to find remnant mass in GiantBranch::CalculateRemnantMassByMullerMandel()
+constexpr int    MULLERMANDEL_REMNANT_MASS_MAX_ITERATIONS = 1000;                                                   // Maximum number of iterations to find remnant mass in GiantBranch::CalculateRemnantMassByMullerMandel()
 
 constexpr int    PULSAR_SPIN_ITERATIONS                 = 100;                                                      // Maximum number of iterations to find pulsar birth spin period in NS::CalculatePulsarBirthSpinPeriod()